[Refactor - #56] 예외 클래스 구분을 위한 패키지 생성 및 이동
