+++ src/main/java/com/dateplan/dateplan/domain/member/service/MemberReadService.java
@@ -4,8 +4,8 @@ import com.dateplan.dateplan.domain.member.dto.MemberInfoServiceResponse;
 import com.dateplan.dateplan.domain.member.entity.Member;
 import com.dateplan.dateplan.domain.member.repository.MemberRepository;
 import com.dateplan.dateplan.global.auth.MemberThreadLocal;
-import com.dateplan.dateplan.global.exception.AlReadyRegisteredNicknameException;
-import com.dateplan.dateplan.global.exception.AlReadyRegisteredPhoneException;
+import com.dateplan.dateplan.global.exception.member.AlReadyRegisteredNicknameException;
+import com.dateplan.dateplan.global.exception.member.AlReadyRegisteredPhoneException;
 import com.dateplan.dateplan.global.exception.auth.MemberNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
