+++ src/main/java/com/dateplan/dateplan/global/exception/auth/NoPermissionException.java
@@ -1,7 +1,9 @@
-package com.dateplan.dateplan.global.exception;
+package com.dateplan.dateplan.global.exception.auth;
 
 import com.dateplan.dateplan.global.constant.Operation;
 import com.dateplan.dateplan.global.constant.Resource;
+import com.dateplan.dateplan.global.exception.ApplicationException;
+import com.dateplan.dateplan.global.exception.ErrorCode;
 import com.dateplan.dateplan.global.exception.ErrorCode.DetailMessage;
 import java.io.Serial;
