+++ src/main/java/com/dateplan/dateplan/global/exception/member/AlReadyRegisteredPhoneException.java
@@ -1,5 +1,7 @@
-package com.dateplan.dateplan.global.exception;
+package com.dateplan.dateplan.global.exception.member;
 
+import com.dateplan.dateplan.global.exception.ApplicationException;
+import com.dateplan.dateplan.global.exception.ErrorCode;
 import com.dateplan.dateplan.global.exception.ErrorCode.DetailMessage;
 import java.io.Serial;
