+++ src/test/java/com/dateplan/dateplan/controller/member/AuthController/AuthControllerTest.java
@@ -31,11 +31,11 @@ import com.dateplan.dateplan.domain.member.dto.signup.SignUpRequest;
 import com.dateplan.dateplan.domain.member.dto.signup.SignUpServiceRequest;
 import com.dateplan.dateplan.domain.sms.type.SmsType;
 import com.dateplan.dateplan.global.constant.Gender;
-import com.dateplan.dateplan.global.exception.AlReadyRegisteredNicknameException;
-import com.dateplan.dateplan.global.exception.AlReadyRegisteredPhoneException;
+import com.dateplan.dateplan.global.exception.member.AlReadyRegisteredNicknameException;
+import com.dateplan.dateplan.global.exception.member.AlReadyRegisteredPhoneException;
 import com.dateplan.dateplan.global.exception.ErrorCode.DetailMessage;
-import com.dateplan.dateplan.global.exception.InvalidPhoneAuthCodeException;
-import com.dateplan.dateplan.global.exception.PhoneNotAuthenticatedException;
+import com.dateplan.dateplan.global.exception.auth.InvalidPhoneAuthCodeException;
+import com.dateplan.dateplan.global.exception.auth.PhoneNotAuthenticatedException;
 import com.dateplan.dateplan.global.exception.auth.MemberNotFoundException;
 import com.dateplan.dateplan.global.exception.auth.PasswordMismatchException;
 import com.dateplan.dateplan.global.exception.auth.TokenInvalidException;
