+++ src/test/java/com/dateplan/dateplan/service/member/AuthServiceTest.java
@@ -26,9 +26,9 @@ import com.dateplan.dateplan.domain.member.repository.MemberRepository;
 import com.dateplan.dateplan.domain.member.service.AuthService;
 import com.dateplan.dateplan.domain.sms.type.SmsType;
 import com.dateplan.dateplan.global.constant.Gender;
-import com.dateplan.dateplan.global.exception.AlReadyRegisteredPhoneException;
-import com.dateplan.dateplan.global.exception.InvalidPhoneAuthCodeException;
-import com.dateplan.dateplan.global.exception.PhoneNotAuthenticatedException;
+import com.dateplan.dateplan.global.exception.member.AlReadyRegisteredPhoneException;
+import com.dateplan.dateplan.global.exception.auth.InvalidPhoneAuthCodeException;
+import com.dateplan.dateplan.global.exception.auth.PhoneNotAuthenticatedException;
 import com.dateplan.dateplan.global.exception.auth.PasswordMismatchException;
 import com.dateplan.dateplan.global.exception.sms.SmsSendFailException;
 import com.dateplan.dateplan.global.util.RandomCodeGenerator;
