+++ src/test/java/com/dateplan/dateplan/service/member/MemberReadServiceTest.java
@@ -10,8 +10,8 @@ import com.dateplan.dateplan.domain.member.repository.MemberRepository;
 import com.dateplan.dateplan.domain.member.service.MemberReadService;
 import com.dateplan.dateplan.global.auth.MemberThreadLocal;
 import com.dateplan.dateplan.global.constant.Gender;
-import com.dateplan.dateplan.global.exception.AlReadyRegisteredNicknameException;
-import com.dateplan.dateplan.global.exception.AlReadyRegisteredPhoneException;
+import com.dateplan.dateplan.global.exception.member.AlReadyRegisteredNicknameException;
+import com.dateplan.dateplan.global.exception.member.AlReadyRegisteredPhoneException;
 import com.dateplan.dateplan.global.exception.ErrorCode.DetailMessage;
 import com.dateplan.dateplan.global.exception.auth.MemberNotFoundException;
 import com.dateplan.dateplan.service.ServiceTestSupport;
