+++ src/test/java/com/dateplan/dateplan/service/member/MemberServiceTest.java
@@ -31,10 +31,10 @@ import com.dateplan.dateplan.global.auth.MemberThreadLocal;
 import com.dateplan.dateplan.global.constant.Gender;
 import com.dateplan.dateplan.global.constant.Operation;
 import com.dateplan.dateplan.global.constant.Resource;
-import com.dateplan.dateplan.global.exception.AlReadyRegisteredNicknameException;
-import com.dateplan.dateplan.global.exception.AlReadyRegisteredPhoneException;
-import com.dateplan.dateplan.global.exception.NoPermissionException;
-import com.dateplan.dateplan.global.exception.PhoneNotAuthenticatedException;
+import com.dateplan.dateplan.global.exception.member.AlReadyRegisteredNicknameException;
+import com.dateplan.dateplan.global.exception.member.AlReadyRegisteredPhoneException;
+import com.dateplan.dateplan.global.exception.auth.NoPermissionException;
+import com.dateplan.dateplan.global.exception.auth.PhoneNotAuthenticatedException;
 import com.dateplan.dateplan.global.exception.S3Exception;
 import com.dateplan.dateplan.global.exception.S3ImageNotFoundException;
 import com.dateplan.dateplan.service.ServiceTestSupport;
